ageGroupChild

+++ registration-processor/pre-processor/registration-processor-packet-classifier-stage/src/main/java/io/mosip/registration/processor/stages/packetclassifier/tagging/impl/AgeGroupTagGenerator.java
@@ -42,16 +42,16 @@ public class AgeGroupTagGenerator implements TagGenerator {
 
     private static String RANGE_DELIMITER = "-";
 
-    private Map<String, int[]> parsedAgeGroupRangemap;
+    private Map<String, double[]> parsedAgeGroupRangemap;
 
     @PostConstruct
     private void generateParsedAgeGroupRangeMap() {
         parsedAgeGroupRangemap = new HashMap<>();
         for (Map.Entry<String,String> entry : ageGroupRangeMap.entrySet()) {
             String[] range = entry.getValue().split(RANGE_DELIMITER);
-            int[] rangeArray = new int[2];
-            rangeArray[0] = Integer.parseInt(range[0]);
-            rangeArray[1] = Integer.parseInt(range[1]);
+            double[] rangeArray = new double[2];
+            rangeArray[0] = Double.parseDouble(range[0]);
+            rangeArray[1] = Double.parseDouble(range[1]);
             parsedAgeGroupRangemap.put(entry.getKey(), rangeArray);
         }
     }
@@ -73,12 +73,12 @@ public class AgeGroupTagGenerator implements TagGenerator {
                 throws BaseCheckedException {
         try {
             String ageGroup = "";
-            int age = utility.getApplicantAge(registrationId, process, ProviderStageName.CLASSIFICATION);
+            double age = utility.getApplicantAge(registrationId, process, ProviderStageName.CLASSIFICATION);
             
 			if (age == -1) {
 				ageGroup = notAvailableTagValue;
 			} else {
-				for (Map.Entry<String, int[]> entry : parsedAgeGroupRangemap.entrySet()) {
+				for (Map.Entry<String, double[]> entry : parsedAgeGroupRangemap.entrySet()) {
 					if (age >= entry.getValue()[0] && age <= entry.getValue()[1]) {
 						ageGroup = entry.getKey();
 						break;
